Norms can be read from file and (de)activated when specified. Instead of a number of iterations, the program now also accepts an end-date as an argument

+++ src/main/java/nl/uu/iss/ga/Simulation.java
@@ -96,7 +96,7 @@ public class Simulation {
     }
 
     private SimulationEngine<CandidateActivity> getLocalSimulationEngine() {
-        return new DefaultSimulationEngine<>(this.platform, arguments.getIterations(), this.environmentInterface);
+        return new DefaultSimulationEngine<>(this.platform, (int)arguments.getIterations(), this.environmentInterface);
     }
 
     private SimulationEngine<CandidateActivity> getPansimSimulationEngine() {
